refactored: move PersonInput to resolvers and make dob optional
